+++ apps/frontend/src/libs/components/header/header.tsx
@@ -1,5 +1,5 @@
 import defaultAvatar from "~/assets/img/default-avatar.png";
-import { Button, Icon } from "~/libs/components/components.js";
+import { Button } from "~/libs/components/components.js";
 import { useAppSelector } from "~/libs/hooks/hooks.js";
 import { type UserDto } from "~/modules/users/users.js";
 
@@ -16,9 +16,10 @@ const Header: React.FC<Properties> = ({ onSidebarToggle }: Properties) => {
 		<header className={styles["header"]}>
 			<div className={styles["menu-btn"]}>
 				<Button
-					icon={<Icon name="menu" />}
+					hasVisuallyHiddenLabel
+					iconName="menu"
+					label="Menu"
 					onClick={onSidebarToggle}
-					variant="secondary"
 				/>
 			</div>