+++ apps/frontend/src/libs/components/icon/libs/maps/icon-name-to-svg.map.ts
@@ -1,4 +1,6 @@
 import Close from "~/assets/img/close.svg?react";
+import CrossedEye from "~/assets/img/crossed-eye.svg?react";
+import Eye from "~/assets/img/eye.svg?react";
 import Menu from "~/assets/img/menu.svg?react";
 import WheelOfBalanceActive from "~/assets/img/wheel-of-balance-active.svg?react";
 import WheelOfBalanceInactive from "~/assets/img/wheel-of-balance-inactive.svg?react";
@@ -9,6 +11,8 @@ const iconNameToSvg: Record<
 	React.FC<React.SVGProps<SVGSVGElement>>
 > = {
 	close: Close,
+	crossedEye: CrossedEye,
+	eye: Eye,
 	menu: Menu,
 	wheelActive: WheelOfBalanceActive,
 	wheelInactive: WheelOfBalanceInactive,