+++ apps/frontend/src/pages/auth/libs/components/sign-up-form/libs/constants.ts
@@ -1,6 +1,7 @@
-import { type UserSignUpRequestDto } from "~/modules/users/users.js";
+import { type UserSignUpFormDto } from "~/modules/users/users.js";
 
-const DEFAULT_SIGN_UP_PAYLOAD: UserSignUpRequestDto = {
+const DEFAULT_SIGN_UP_PAYLOAD: UserSignUpFormDto = {
+	confirmPassword: "",
 	email: "",
 	name: "",
 	password: "",